+++ PostHog/PostHogStorage.swift
@@ -242,6 +242,7 @@ class PostHogStorage {
         case remoteConfig = "posthog.remoteConfig"
         case surveySeen = "posthog.surveySeen"
         case requestId = "posthog.requestId"
+        case evaluatedAt = "posthog.evaluatedAt"
         case personPropertiesForFlags = "posthog.personPropertiesForFlags"
         case groupPropertiesForFlags = "posthog.groupPropertiesForFlags"
     }
@@ -444,6 +445,28 @@ class PostHogStorage {
     }
 
     func setBool(forKey key: StorageKey, contents: Bool) {
+        setJson(forKey: key, json: contents)
+    }
+
+    func getInt(forKey key: StorageKey) -> Int? {
+        let value = getJson(forKey: key)
+        if let intValue = value as? Int {
+            return intValue
+        } else if let numberValue = value as? NSNumber {
+            return numberValue.intValue
+        } else if let dictValue = value as? [String: Any],
+                  let nestedValue = dictValue[key.rawValue]
+        {
+            if let intValue = nestedValue as? Int {
+                return intValue
+            } else if let numberValue = nestedValue as? NSNumber {
+                return numberValue.intValue
+            }
+        }
+        return nil
+    }
+
+    func setInt(forKey key: StorageKey, contents: Int) {
         setJson(forKey: key, json: contents)
     }
 }