+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - feat: add ability to override sendFeatureFlagEvent for exact feature flag call ([#396](https://github.com/PostHog/posthog-ios/pull/396))
 
+## 3.36.0 - 2025-12-08
+
+- feat: include `evaluated_at` properties in `$feature_flag_called` events ([#394](https://github.com/PostHog/posthog-ios/pull/394))
+- fix: only report $feature_flag_called if the flag value has changed ([#405](https://github.com/PostHog/posthog-ios/pull/405))
+
 ## 3.35.1 - 2025-12-02
 
 - fix: avoid memory leaks on foat conversions ([#401](https://github.com/PostHog/posthog-ios/pull/401))